+++ wfsim/__init__.py
@@ -1,5 +1,13 @@
-__version__ = "0.5.5"
+__version__ = "0.5.6"
 
-from .core import *
-from .strax_interface import *
+from .core.afterpulse import *
+from .core.pulse import *
+from .core.rawdata import *
+from .core.s1 import *
+from .core.s2 import *
+
 from .pax_interface import *
+from .strax_interface import *
+
+from .load_resource import *
+from .utils import *